requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
playsound==1.3.0
PyJWT==2.6.0
python-decouple==3.8
pytz==2023.3
requests==2.28.2
twilio==8.0.0
urllib3==1.26.15
yarl==1.8.2
